requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
aniso8601==9.0.1
annotated-types==0.5.0
anyio==3.7.1
argcomplete==3.2.1
arrow==1.3.0
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bitarray==2.9.2
blinker==1.8.2
boto3==1.35.1
botocore==1.35.1
build==1.0.3
CacheControl==0.14.0
catalogue==2.0.10
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
cohere==5.8.1
colbert-ai==0.2.19
colorama==0.4.6
colorlog==6.8.2
comm==0.1.3
contourpy==1.2.1
coverage==7.6.1
crashtest==0.4.1
customtkinter==5.2.1
cycler==0.12.1
darkdetect==0.8.0
dataclasses-json==0.6.7
datasets==2.20.0
debugpy==1.6.7
decorator==5.1.1
deepdiff==7.0.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distlib==0.3.8
distro==1.8.0
dnspython==2.4.2
docker==7.0.0
dulwich==0.21.7
emoji==2.12.1
et-xmlfile==1.1.0
executing==1.2.0
faiss-cpu==1.8.0.post1
fast-pytorch-kmeans==0.2.0.1
fastapi==0.103.2
fastavro==1.9.5
fastjsonschema==2.19.0
filelock==3.13.1
filetype==1.2.0
flasgger==0.9.7.1
Flask==3.0.3
Flask-Alembic==3.1.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
fonttools==4.53.1
frozenlist==1.4.0
fsspec==2024.5.0
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.43
googleapis-common-protos==1.63.2
greenlet==3.0.1
grobid-client-python==0.0.8
grpcio==1.65.0
h11==0.14.0
httpcore==1.0.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.5
idna==3.4
importlib-metadata==7.0.1
importlib_resources==6.4.4
iniconfig==2.0.0
installer==0.7.0
intel-openmp==2021.4.0
ipykernel==6.23.1
ipython==8.14.0
itsdangerous==2.2.0
jaraco.classes==3.3.0
jedi==0.18.2
Jinja2==3.1.4
jinja2-time==0.2.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.2.0
jupyter_core==5.3.0
keyring==24.3.0
kiwisolver==1.4.5
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-text-splitters==0.0.2
langdetect==1.0.9
langsmith==0.1.88
llama-cloud==0.0.9
llama-index==0.10.55
llama-index-agent-openai==0.2.8
llama-index-cli==0.1.12
llama-index-core==0.10.55
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.2.5
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.25
llama-index-multi-modal-llms-openai==0.1.7
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.30
llama-index-readers-llama-parse==0.1.6
llama-parse==0.4.9
loguru==0.7.2
lxml==5.2.2
lz4==4.3.3
make==0.1.6.post2
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.20.1
matplotlib==3.9.1
matplotlib-inline==0.1.6
matrix-py3==0.0.4
mistune==3.0.2
mkl==2021.4.0
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nltk==3.8.1
numpy==1.26.1
onnx==1.16.1
openai==1.35.10
openpyxl==3.1.2
ordered-set==4.1.0
orjson==3.10.6
packaging==23.2
pandas==2.2.2
parameterized==0.9.0
parso==0.8.3
Pebble==5.0.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pinecone==4.0.0
pinecone-client==2.2.4
pipx==1.3.3
pkginfo==1.9.6
platformdirs==3.11.0
pluggy==1.5.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.6.0
prompt-toolkit==3.0.38
protobuf==4.25.3
protoc-gen-openapiv2==0.0.1
psutil==5.9.5
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==17.0.0
pyarrow-hotfix==0.6
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.15.1
PyJWT==2.9.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pynvml==11.5.3
pyparsing==3.1.2
pypdf==4.3.0
pyproject_hooks==1.0.0
pytest==8.3.2
pytest-cov==5.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-slugify==8.0.4
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.1.0
RAGatouille==0.0.8.post2
rapidfuzz==3.5.2
referencing==0.35.1
regex==2023.12.25
requests==2.32.3
requests-mock==1.12.1
requests-toolbelt==1.0.0
rpds-py==0.20.0
s3transfer==0.10.2
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
semantic-chunkers==0.0.9
semantic-router==0.0.60
sentence-transformers==2.7.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
srsly==2.4.8
stack-data==0.6.2
starlette==0.27.0
striprtf==0.0.26
swagger_ui_bundle==1.1.0
sympy==1.13.0
tabulate==0.9.0
tbb==2021.13.0
tenacity==8.2.3
text-unidecode==1.3
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomlkit==0.12.3
torch==2.3.1
tornado==6.3.2
tqdm==4.66.4
traitlets==5.9.0
transformers==4.42.4
trove-classifiers==2023.11.29
types-python-dateutil==2.9.0.20240316
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2024.1
ujson==5.10.0
unstructured==0.15.1
unstructured-client==0.25.4
urllib3==2.0.7
userpath==1.9.1
uvicorn==0.23.2
virtualenv==20.25.0
voyager==2.0.7
wcwidth==0.2.6
Werkzeug==3.0.3
win32-setctime==1.1.0
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
